Add initial support for EBML
These are the first steps towards MKV native support and solving issues
related to Hls live discontinuities.

https://www.matroska.org/technical/streaming.html

"Another option is to concatenate Segment Elements with known sizes,
one after the other. This solution allows a change of codec/resolution
between each segment. For example, this allows for a switch between 4:3
and 16:9 in a television program."